requirements.txt
contourpy
cycler
fabio
fonttools
h5py
matplotlib
numexpr!=2.8.6
numpy
packaging
pandas
Pillow
pyfai
pygix
pyparsing
PyQt5
PyQt5-Qt5
PyQt5-sip
python-dateutil
pytz
scipy
silx
six
tzdata
meson
ninja
pytest
pydantic
watchdog